Move metadata fetch methods to global class
