+++ src/gtdpad/api/infrastructure/Global.cs
@@ -1,5 +1,8 @@
 using System.Net;
 using System.Net.Http;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using HtmlAgilityPack;
 
 namespace gtdpad
 {
@@ -18,6 +21,83 @@ namespace gtdpad
                 "User-Agent", 
                 "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.99 Safari/537.36"
             );
+        }
+
+        private static async Task<string> FetchMetadata(string url)
+        {
+            try
+            {
+                return await Global.HttpClient.GetStringAsync(url);
+            }
+            catch
+            {
+                return null;
+            }
+        }
+
+        public static Metadata ExtractMetaData(string rqurl)
+        {
+            var content = FetchMetadata(rqurl).Result;
+            
+            if(!string.IsNullOrWhiteSpace(content))
+            {
+                var html = new HtmlDocument();
+                html.LoadHtml(content);
+
+                var data = new Metadata {
+                    // Url = rqurl,
+                    Title = html.DocumentNode.SelectSingleNode("//title")?.InnerText
+                };
+
+                var titleTags = new List<string> { 
+                    "//meta[@property='og:title']", 
+                    "//meta[@property='twitter:title']" 
+                };
+
+                var descriptionTags = new List<string> {
+                    "//meta[@name='description']",
+                    "//meta[@property='og:description']",
+                    "//meta[@property='twitter:description']"
+                };
+
+                var imageTags = new List<string> {
+                    "//meta[@property='og:image']",
+                    "//meta[@property='twitter:image']"
+                };
+
+                var urlTags = new List<string> {
+                    "//meta[@property='og:url']",
+                    "//meta[@property='twitter:url']"
+                };
+
+                titleTags.ForEach(xpath => {
+                    var title = html.GetText(xpath);
+                    if(!string.IsNullOrWhiteSpace(title))
+                        data.Title = title;
+                });
+
+                // descriptionTags.ForEach(xpath => {
+                //     var description = html.GetText(xpath);
+                //     if(!string.IsNullOrWhiteSpace(description))
+                //         data.Description = description;
+                // });
+
+                // imageTags.ForEach(xpath => {
+                //     var image = html.GetText(xpath);
+                //     if(!string.IsNullOrWhiteSpace(image))
+                //         data.Image = image;
+                // });
+
+                // urlTags.ForEach(xpath => {
+                //     var url = html.GetText(xpath);
+                //     if(!string.IsNullOrWhiteSpace(url))
+                //         data.Url = url;
+                // });
+
+                return data;
+            }
+
+            return null;
         }
     }
 }
